Release v3.0: Implementado padrão Singleton, correção de layout na view do instrutor e documentação atualizada

+++ dao/aluno_dao.py
@@ -1,21 +1,24 @@
-from config.database import Database
+from config.database import DatabaseConnection
 from model.aluno import Aluno
 
 class AlunoDAO:
+    def __init__(self):
+        # Pega a conexão única compartilhada pelo Singleton
+        self.conn = DatabaseConnection()
+
     def cadastrar(self, aluno: Aluno):
-        conn = Database.get_connection()
-        cursor = conn.cursor()
+        if not self.conn: return None
+        cursor = self.conn.cursor()
         query = "INSERT INTO aluno (nome, cpf, telefone, id_instrutor) VALUES (%s, %s, %s, %s) RETURNING id"
         cursor.execute(query, (aluno.nome, aluno.cpf, aluno.telefone, aluno.id_instrutor))
         aluno.id = cursor.fetchone()[0]
-        conn.commit()
+        self.conn.commit()
         cursor.close()
-        conn.close()
         return aluno.id
 
     def listar_todos(self, busca=""):
-        conn = Database.get_connection()
-        cursor = conn.cursor()
+        if not self.conn: return []
+        cursor = self.conn.cursor()
         query = """SELECT a.id, a.nome, a.cpf, a.telefone, a.id_instrutor, i.nome 
                    FROM aluno a 
                    LEFT JOIN instrutor i ON a.id_instrutor = i.id"""
@@ -29,32 +32,28 @@ class AlunoDAO:
         rows = cursor.fetchall()
         alunos = [Aluno(r[0], r[1], r[2], r[3], r[4], r[5]) for r in rows]
         cursor.close()
-        conn.close()
         return alunos
 
     def atualizar(self, aluno: Aluno):
-        conn = Database.get_connection()
-        cursor = conn.cursor()
+        if not self.conn: return
+        cursor = self.conn.cursor()
         query = "UPDATE aluno SET nome=%s, cpf=%s, telefone=%s, id_instrutor=%s WHERE id=%s"
         cursor.execute(query, (aluno.nome, aluno.cpf, aluno.telefone, aluno.id_instrutor, aluno.id))
-        conn.commit()
+        self.conn.commit()
         cursor.close()
-        conn.close()
 
     def deletar(self, id_aluno):
-        conn = Database.get_connection()
-        cursor = conn.cursor()
+        if not self.conn: return
+        cursor = self.conn.cursor()
         query = "DELETE FROM aluno WHERE id = %s"
         cursor.execute(query, (id_aluno,))
-        conn.commit()
+        self.conn.commit()
         cursor.close()
-        conn.close()
 
     def listar_instrutores(self):
-        conn = Database.get_connection()
-        cursor = conn.cursor()
+        if not self.conn: return []
+        cursor = self.conn.cursor()
         cursor.execute("SELECT id, nome FROM instrutor ORDER BY nome")
         rows = cursor.fetchall()
         cursor.close()
-        conn.close()
         return rows
